Generate code with swagger-codegen 2.x
The swagger-codegen 3.x version currently generates type mappings to
`BigDecimal` for numbers, which isn't a native PHP class.

+++ lib/Model/PersonAutoCompleteDTO.php
@@ -15,11 +15,12 @@
  *
  * The Tripletex API is a **RESTful API**, which does not implement PATCH, but uses a PUT with optional fields.  **Actions** or commands are represented in our RESTful path with a prefixed `:`. Example: `/v2/hours/123/:approve`.  **Summaries** or aggregated results are represented in our RESTful path with a prefixed <code>&gt;</code>. Example: <code>/v2/hours/&gt;thisWeeksBillables</code>.  **\"requestID\"** is a key found in all validation and error responses. If additional log information is absolutely necessary, our support division can locate the key value.  **Download** the [swagger.json](/v2/swagger.json) file [OpenAPI Specification](https://github.com/OAI/OpenAPI-Specification) to [generate code](https://github.com/swagger-api/swagger-codegen). This document was generated from the Swagger JSON file.  **version:** This is a versioning number found on all DB records. If included, it will prevent your PUT/POST from overriding any updates to the record since your GET.  **Date & DateTime** follows the **ISO 8601** standard. Date: `YYYY-MM-DD`. DateTime: `YYYY-MM-DDThh:mm:ssZ`  **Sorting** is done by specifying a comma separated list, where a `-` prefix denotes descending. You can sort by sub object with the following format: `project.name, -date`.  **Searching:** is done by entering values in the optional fields for each API call. The values fall into the following categories: range, in, exact and like.  **Missing fields or even no response data** can occur because result objects and fields are filtered on authorization.  **See [FAQ](https://tripletex.no/execute/docViewer?articleId=906&language=0) for more additional information.**   ## Authentication: - **Tokens:** The Tripletex API uses 3 different tokens - **consumerToken**, **employeeToken** and **sessionToken**.  - **consumerToken** is a token provided to the consumer by Tripletex after the API 2.0 registration is completed.  - **employeeToken** is a token created by an administrator in your Tripletex account via the user settings and the tab \"API access\". Each employee token must be given a set of entitlements. [Read more here.](https://tripletex.no/execute/docViewer?articleId=853&language=0)  - **sessionToken** is the token from `/token/session/:create` which requires a consumerToken and an employeeToken created with the same consumer token, but not an authentication header. See how to create a sessionToken [here](https://tripletex.no/execute/docViewer?articleId=855&language=0). - The session token is used as the password in \"Basic Authentication Header\" for API calls.  - Use blank or `0` as username for accessing the account with regular employee token, or if a company owned employee token accesses <code>/company/&gt;withLoginAccess</code> or <code>/token/session/&gt;whoAmI</code>.  - For company owned employee tokens (accounting offices) the ID from <code>/company/&gt;withLoginAccess</code> can be used as username for accessing client accounts.  - If you need to create the header yourself use <code>Authorization: Basic &lt;base64encode('0:sessionToken')&gt;</code>.   ## Tags: - <div class=\"tag-icon-beta\"></div> **[BETA]** This is a beta endpoint and can be subject to change. - <div class=\"tag-icon-deprecated\"></div> **[DEPRECATED]** Deprecated means that we intend to remove/change this feature or capability in a future \"major\" API release. We therefore discourage all use of this feature/capability.  ## Fields: Use the `fields` parameter to specify which fields should be returned. This also supports fields from sub elements. Example values: - `project,activity,hours`  returns `{project:..., activity:...., hours:...}`. - just `project` returns `\"project\" : { \"id\": 12345, \"url\": \"tripletex.no/v2/projects/12345\"  }`. - `project(*)` returns `\"project\" : { \"id\": 12345 \"name\":\"ProjectName\" \"number.....startDate\": \"2013-01-07\" }`. - `project(name)` returns `\"project\" : { \"name\":\"ProjectName\" }`. - All elements and some subElements :  `*,activity(name),employee(*)`.  ## Changes: To get the changes for a resource, `changes` have to be explicitly specified as part of the `fields` parameter, e.g. `*,changes`. There are currently two types of change available:  - `CREATE` for when the resource was created - `UPDATE` for when the resource was updated  NOTE: For objects created prior to October 24th 2018 the list may be incomplete, but will always contain the CREATE and the last change (if the object has been changed after creation).  ## Rate limiting in each response header: Rate limiting is performed on the API calls for an employee for each API consumer. Status regarding the rate limit is returned as headers: - `X-Rate-Limit-Limit` - The number of allowed requests in the current period. - `X-Rate-Limit-Remaining` - The number of remaining requests. - `X-Rate-Limit-Reset` - The number of seconds left in the current period.  Once the rate limit is hit, all requests will return HTTP status code `429` for the remainder of the current period.   ## Response envelope: ``` {   \"fullResultSize\": ###,   \"from\": ###, // Paging starting from   \"count\": ###, // Paging count   \"versionDigest\": \"Hash of full result\",   \"values\": [...list of objects...] } {   \"value\": {...single object...} } ```   ## WebHook envelope: ``` {   \"subscriptionId\": ###,   \"event\": \"object.verb\", // As listed from /v2/event/   \"id\": ###, // Object id   \"value\": {... single object, null if object.deleted ...} } ```    ## Error/warning envelope: ``` {   \"status\": ###, // HTTP status code   \"code\": #####, // internal status code of event   \"message\": \"Basic feedback message in your language\",   \"link\": \"Link to doc\",   \"developerMessage\": \"More technical message\",   \"validationMessages\": [ // Will be null if Error     {       \"field\": \"Name of field\",       \"message\": \"Validation failure information\"     }   ],   \"requestId\": \"UUID used in any logs\" } ```   ## Status codes / Error codes: - **200 OK** - **201 Created** - From POSTs that create something new. - **204 No Content** - When there is no answer, ex: \"/:anAction\" or DELETE. - **400 Bad request** -   - **4000** Bad Request Exception   - **11000** Illegal Filter Exception   - **12000** Path Param Exception   - **24000**   Cryptography Exception - **401 Unauthorized** - When authentication is required and has failed or has not yet been provided   -  **3000** Authentication Exception   -  **9000** Security Exception - **403 Forbidden** - When AuthorisationManager says no. - **404 Not Found** - For content/IDs that does not exist.   -  **6000** Not Found Exception - **409 Conflict** - Such as an edit conflict between multiple simultaneous updates   -  **7000** Object Exists Exception   -  **8000** Revision Exception   - **10000** Locked Exception   - **14000** Duplicate entry - **422 Bad Request** - For Required fields or things like malformed payload.   - **15000** Value Validation Exception   - **16000** Mapping Exception   - **17000** Sorting Exception   - **18000** Validation Exception   - **21000** Param Exception   - **22000** Invalid JSON Exception   - **23000**   Result Set Too Large Exception - **429 Too Many Requests** - Request rate limit hit - **500 Internal Error** -  Unexpected condition was encountered and no more specific message is suitable   -  **1000** Exception
  *
- * OpenAPI spec version: 2.33.1
+ * OpenAPI spec version: 2.33.2
  * 
  * Generated by: https://github.com/swagger-api/swagger-codegen.git
- * Swagger Codegen version: 3.0.5
+ * Swagger Codegen version: 2.4.2
  */
+
 /**
  * NOTE: This class is auto generated by the swagger code generator program.
  * https://github.com/swagger-api/swagger-codegen
@@ -57,22 +58,23 @@ class PersonAutoCompleteDTO implements ModelInterface, ArrayAccess
       */
     protected static $swaggerTypes = [
         'name' => 'string',
-'alternate_name' => 'string',
-'address' => 'string',
-'postal_code' => 'string',
-'postal_area' => 'string',
-'latitude' => 'BigDecimal',
-'longitude' => 'BigDecimal',
-'score' => 'int',
-'sources' => 'string[]',
-'firstname' => 'string',
-'lastname' => 'string',
-'phone_number' => 'string',
-'phone_number_mobile' => 'string',
-'email' => 'string',
-'url' => 'string',
-'country_id' => 'int',
-'reserved' => 'bool'    ];
+        'alternate_name' => 'string',
+        'address' => 'string',
+        'postal_code' => 'string',
+        'postal_area' => 'string',
+        'latitude' => 'float',
+        'longitude' => 'float',
+        'score' => 'int',
+        'sources' => 'string[]',
+        'firstname' => 'string',
+        'lastname' => 'string',
+        'phone_number' => 'string',
+        'phone_number_mobile' => 'string',
+        'email' => 'string',
+        'url' => 'string',
+        'country_id' => 'int',
+        'reserved' => 'bool'
+    ];
 
     /**
       * Array of property to format mappings. Used for (de)serialization
@@ -81,22 +83,23 @@ class PersonAutoCompleteDTO implements ModelInterface, ArrayAccess
       */
     protected static $swaggerFormats = [
         'name' => null,
-'alternate_name' => null,
-'address' => null,
-'postal_code' => null,
-'postal_area' => null,
-'latitude' => null,
-'longitude' => null,
-'score' => 'int32',
-'sources' => null,
-'firstname' => null,
-'lastname' => null,
-'phone_number' => null,
-'phone_number_mobile' => null,
-'email' => null,
-'url' => null,
-'country_id' => 'int32',
-'reserved' => null    ];
+        'alternate_name' => null,
+        'address' => null,
+        'postal_code' => null,
+        'postal_area' => null,
+        'latitude' => null,
+        'longitude' => null,
+        'score' => 'int32',
+        'sources' => null,
+        'firstname' => null,
+        'lastname' => null,
+        'phone_number' => null,
+        'phone_number_mobile' => null,
+        'email' => null,
+        'url' => null,
+        'country_id' => 'int32',
+        'reserved' => null
+    ];
 
     /**
      * Array of property to type mappings. Used for (de)serialization
@@ -126,22 +129,23 @@ class PersonAutoCompleteDTO implements ModelInterface, ArrayAccess
      */
     protected static $attributeMap = [
         'name' => 'name',
-'alternate_name' => 'alternateName',
-'address' => 'address',
-'postal_code' => 'postalCode',
-'postal_area' => 'postalArea',
-'latitude' => 'latitude',
-'longitude' => 'longitude',
-'score' => 'score',
-'sources' => 'sources',
-'firstname' => 'firstname',
-'lastname' => 'lastname',
-'phone_number' => 'phoneNumber',
-'phone_number_mobile' => 'phoneNumberMobile',
-'email' => 'email',
-'url' => 'url',
-'country_id' => 'countryId',
-'reserved' => 'reserved'    ];
+        'alternate_name' => 'alternateName',
+        'address' => 'address',
+        'postal_code' => 'postalCode',
+        'postal_area' => 'postalArea',
+        'latitude' => 'latitude',
+        'longitude' => 'longitude',
+        'score' => 'score',
+        'sources' => 'sources',
+        'firstname' => 'firstname',
+        'lastname' => 'lastname',
+        'phone_number' => 'phoneNumber',
+        'phone_number_mobile' => 'phoneNumberMobile',
+        'email' => 'email',
+        'url' => 'url',
+        'country_id' => 'countryId',
+        'reserved' => 'reserved'
+    ];
 
     /**
      * Array of attributes to setter functions (for deserialization of responses)
@@ -150,22 +154,23 @@ class PersonAutoCompleteDTO implements ModelInterface, ArrayAccess
      */
     protected static $setters = [
         'name' => 'setName',
-'alternate_name' => 'setAlternateName',
-'address' => 'setAddress',
-'postal_code' => 'setPostalCode',
-'postal_area' => 'setPostalArea',
-'latitude' => 'setLatitude',
-'longitude' => 'setLongitude',
-'score' => 'setScore',
-'sources' => 'setSources',
-'firstname' => 'setFirstname',
-'lastname' => 'setLastname',
-'phone_number' => 'setPhoneNumber',
-'phone_number_mobile' => 'setPhoneNumberMobile',
-'email' => 'setEmail',
-'url' => 'setUrl',
-'country_id' => 'setCountryId',
-'reserved' => 'setReserved'    ];
+        'alternate_name' => 'setAlternateName',
+        'address' => 'setAddress',
+        'postal_code' => 'setPostalCode',
+        'postal_area' => 'setPostalArea',
+        'latitude' => 'setLatitude',
+        'longitude' => 'setLongitude',
+        'score' => 'setScore',
+        'sources' => 'setSources',
+        'firstname' => 'setFirstname',
+        'lastname' => 'setLastname',
+        'phone_number' => 'setPhoneNumber',
+        'phone_number_mobile' => 'setPhoneNumberMobile',
+        'email' => 'setEmail',
+        'url' => 'setUrl',
+        'country_id' => 'setCountryId',
+        'reserved' => 'setReserved'
+    ];
 
     /**
      * Array of attributes to getter functions (for serialization of requests)
@@ -174,22 +179,23 @@ class PersonAutoCompleteDTO implements ModelInterface, ArrayAccess
      */
     protected static $getters = [
         'name' => 'getName',
-'alternate_name' => 'getAlternateName',
-'address' => 'getAddress',
-'postal_code' => 'getPostalCode',
-'postal_area' => 'getPostalArea',
-'latitude' => 'getLatitude',
-'longitude' => 'getLongitude',
-'score' => 'getScore',
-'sources' => 'getSources',
-'firstname' => 'getFirstname',
-'lastname' => 'getLastname',
-'phone_number' => 'getPhoneNumber',
-'phone_number_mobile' => 'getPhoneNumberMobile',
-'email' => 'getEmail',
-'url' => 'getUrl',
-'country_id' => 'getCountryId',
-'reserved' => 'getReserved'    ];
+        'alternate_name' => 'getAlternateName',
+        'address' => 'getAddress',
+        'postal_code' => 'getPostalCode',
+        'postal_area' => 'getPostalArea',
+        'latitude' => 'getLatitude',
+        'longitude' => 'getLongitude',
+        'score' => 'getScore',
+        'sources' => 'getSources',
+        'firstname' => 'getFirstname',
+        'lastname' => 'getLastname',
+        'phone_number' => 'getPhoneNumber',
+        'phone_number_mobile' => 'getPhoneNumberMobile',
+        'email' => 'getEmail',
+        'url' => 'getUrl',
+        'country_id' => 'getCountryId',
+        'reserved' => 'getReserved'
+    ];
 
     /**
      * Array of attributes where the key is the local name,
@@ -233,20 +239,22 @@ class PersonAutoCompleteDTO implements ModelInterface, ArrayAccess
     }
 
     const SOURCES_SEARCH1881 = 'SEARCH1881';
-const SOURCES_TRIPLETEX = 'TRIPLETEX';
-const SOURCES_NICKNAME = 'NICKNAME';
-const SOURCES_EMPLOYEE = 'EMPLOYEE';
-const SOURCES_CONTACT = 'CONTACT';
-const SOURCES_ACTIVITY = 'ACTIVITY';
-const SOURCES_PROJECT = 'PROJECT';
-const SOURCES_ORDER = 'ORDER';
-const SOURCES_OFFER = 'OFFER';
-const SOURCES_CUSTOMER = 'CUSTOMER';
-const SOURCES_COMPANY = 'COMPANY';
-const SOURCES_CONTROLSCHEMA = 'CONTROLSCHEMA';
-const SOURCES_HOUR = 'HOUR';
-const SOURCES_TRAVELEXPENSE = 'TRAVELEXPENSE';
+    const SOURCES_TRIPLETEX = 'TRIPLETEX';
+    const SOURCES_NICKNAME = 'NICKNAME';
+    const SOURCES_EMPLOYEE = 'EMPLOYEE';
+    const SOURCES_CONTACT = 'CONTACT';
+    const SOURCES_ACTIVITY = 'ACTIVITY';
+    const SOURCES_PROJECT = 'PROJECT';
+    const SOURCES_ORDER = 'ORDER';
+    const SOURCES_OFFER = 'OFFER';
+    const SOURCES_CUSTOMER = 'CUSTOMER';
+    const SOURCES_COMPANY = 'COMPANY';
+    const SOURCES_CONTROLSCHEMA = 'CONTROLSCHEMA';
+    const SOURCES_HOUR = 'HOUR';
+    const SOURCES_TRAVELEXPENSE = 'TRAVELEXPENSE';
+    
 
+    
     /**
      * Gets allowable values of the enum
      *
@@ -256,20 +264,22 @@ const SOURCES_TRAVELEXPENSE = 'TRAVELEXPENSE';
     {
         return [
             self::SOURCES_SEARCH1881,
-self::SOURCES_TRIPLETEX,
-self::SOURCES_NICKNAME,
-self::SOURCES_EMPLOYEE,
-self::SOURCES_CONTACT,
-self::SOURCES_ACTIVITY,
-self::SOURCES_PROJECT,
-self::SOURCES_ORDER,
-self::SOURCES_OFFER,
-self::SOURCES_CUSTOMER,
-self::SOURCES_COMPANY,
-self::SOURCES_CONTROLSCHEMA,
-self::SOURCES_HOUR,
-self::SOURCES_TRAVELEXPENSE,        ];
+            self::SOURCES_TRIPLETEX,
+            self::SOURCES_NICKNAME,
+            self::SOURCES_EMPLOYEE,
+            self::SOURCES_CONTACT,
+            self::SOURCES_ACTIVITY,
+            self::SOURCES_PROJECT,
+            self::SOURCES_ORDER,
+            self::SOURCES_OFFER,
+            self::SOURCES_CUSTOMER,
+            self::SOURCES_COMPANY,
+            self::SOURCES_CONTROLSCHEMA,
+            self::SOURCES_HOUR,
+            self::SOURCES_TRAVELEXPENSE,
+        ];
     }
+    
 
     /**
      * Associative array for storing property values
@@ -452,7 +462,7 @@ self::SOURCES_TRAVELEXPENSE,        ];
     /**
      * Gets latitude
      *
-     * @return BigDecimal
+     * @return float
      */
     public function getLatitude()
     {
@@ -462,7 +472,7 @@ self::SOURCES_TRAVELEXPENSE,        ];
     /**
      * Sets latitude
      *
-     * @param BigDecimal $latitude latitude
+     * @param float $latitude latitude
      *
      * @return $this
      */
@@ -476,7 +486,7 @@ self::SOURCES_TRAVELEXPENSE,        ];
     /**
      * Gets longitude
      *
-     * @return BigDecimal
+     * @return float
      */
     public function getLongitude()
     {
@@ -486,7 +496,7 @@ self::SOURCES_TRAVELEXPENSE,        ];
     /**
      * Sets longitude
      *
-     * @param BigDecimal $longitude longitude
+     * @param float $longitude longitude
      *
      * @return $this
      */
@@ -815,3 +825,5 @@ self::SOURCES_TRAVELEXPENSE,        ];
         return json_encode(ObjectSerializer::sanitizeForSerialization($this));
     }
 }
+
+
