+++ lib/Model/TimesheetEntry.php
@@ -15,11 +15,12 @@
  *
  * The Tripletex API is a **RESTful API**, which does not implement PATCH, but uses a PUT with optional fields.  **Actions** or commands are represented in our RESTful path with a prefixed `:`. Example: `/v2/hours/123/:approve`.  **Summaries** or aggregated results are represented in our RESTful path with a prefixed <code>&gt;</code>. Example: <code>/v2/hours/&gt;thisWeeksBillables</code>.  **\"requestID\"** is a key found in all validation and error responses. If additional log information is absolutely necessary, our support division can locate the key value.  **Download** the [swagger.json](/v2/swagger.json) file [OpenAPI Specification](https://github.com/OAI/OpenAPI-Specification) to [generate code](https://github.com/swagger-api/swagger-codegen). This document was generated from the Swagger JSON file.  **version:** This is a versioning number found on all DB records. If included, it will prevent your PUT/POST from overriding any updates to the record since your GET.  **Date & DateTime** follows the **ISO 8601** standard. Date: `YYYY-MM-DD`. DateTime: `YYYY-MM-DDThh:mm:ssZ`  **Sorting** is done by specifying a comma separated list, where a `-` prefix denotes descending. You can sort by sub object with the following format: `project.name, -date`.  **Searching:** is done by entering values in the optional fields for each API call. The values fall into the following categories: range, in, exact and like.  **Missing fields or even no response data** can occur because result objects and fields are filtered on authorization.  **See [FAQ](https://tripletex.no/execute/docViewer?articleId=906&language=0) for more additional information.**   ## Authentication: - **Tokens:** The Tripletex API uses 3 different tokens - **consumerToken**, **employeeToken** and **sessionToken**.  - **consumerToken** is a token provided to the consumer by Tripletex after the API 2.0 registration is completed.  - **employeeToken** is a token created by an administrator in your Tripletex account via the user settings and the tab \"API access\". Each employee token must be given a set of entitlements. [Read more here.](https://tripletex.no/execute/docViewer?articleId=853&language=0)  - **sessionToken** is the token from `/token/session/:create` which requires a consumerToken and an employeeToken created with the same consumer token, but not an authentication header. See how to create a sessionToken [here](https://tripletex.no/execute/docViewer?articleId=855&language=0). - The session token is used as the password in \"Basic Authentication Header\" for API calls.  - Use blank or `0` as username for accessing the account with regular employee token, or if a company owned employee token accesses <code>/company/&gt;withLoginAccess</code> or <code>/token/session/&gt;whoAmI</code>.  - For company owned employee tokens (accounting offices) the ID from <code>/company/&gt;withLoginAccess</code> can be used as username for accessing client accounts.  - If you need to create the header yourself use <code>Authorization: Basic &lt;base64encode('0:sessionToken')&gt;</code>.   ## Tags: - <div class=\"tag-icon-beta\"></div> **[BETA]** This is a beta endpoint and can be subject to change. - <div class=\"tag-icon-deprecated\"></div> **[DEPRECATED]** Deprecated means that we intend to remove/change this feature or capability in a future \"major\" API release. We therefore discourage all use of this feature/capability.  ## Fields: Use the `fields` parameter to specify which fields should be returned. This also supports fields from sub elements. Example values: - `project,activity,hours`  returns `{project:..., activity:...., hours:...}`. - just `project` returns `\"project\" : { \"id\": 12345, \"url\": \"tripletex.no/v2/projects/12345\"  }`. - `project(*)` returns `\"project\" : { \"id\": 12345 \"name\":\"ProjectName\" \"number.....startDate\": \"2013-01-07\" }`. - `project(name)` returns `\"project\" : { \"name\":\"ProjectName\" }`. - All elements and some subElements :  `*,activity(name),employee(*)`.  ## Changes: To get the changes for a resource, `changes` have to be explicitly specified as part of the `fields` parameter, e.g. `*,changes`. There are currently two types of change available:  - `CREATE` for when the resource was created - `UPDATE` for when the resource was updated  NOTE: For objects created prior to October 24th 2018 the list may be incomplete, but will always contain the CREATE and the last change (if the object has been changed after creation).  ## Rate limiting in each response header: Rate limiting is performed on the API calls for an employee for each API consumer. Status regarding the rate limit is returned as headers: - `X-Rate-Limit-Limit` - The number of allowed requests in the current period. - `X-Rate-Limit-Remaining` - The number of remaining requests. - `X-Rate-Limit-Reset` - The number of seconds left in the current period.  Once the rate limit is hit, all requests will return HTTP status code `429` for the remainder of the current period.   ## Response envelope: ``` {   \"fullResultSize\": ###,   \"from\": ###, // Paging starting from   \"count\": ###, // Paging count   \"versionDigest\": \"Hash of full result\",   \"values\": [...list of objects...] } {   \"value\": {...single object...} } ```   ## WebHook envelope: ``` {   \"subscriptionId\": ###,   \"event\": \"object.verb\", // As listed from /v2/event/   \"id\": ###, // Object id   \"value\": {... single object, null if object.deleted ...} } ```    ## Error/warning envelope: ``` {   \"status\": ###, // HTTP status code   \"code\": #####, // internal status code of event   \"message\": \"Basic feedback message in your language\",   \"link\": \"Link to doc\",   \"developerMessage\": \"More technical message\",   \"validationMessages\": [ // Will be null if Error     {       \"field\": \"Name of field\",       \"message\": \"Validation failure information\"     }   ],   \"requestId\": \"UUID used in any logs\" } ```   ## Status codes / Error codes: - **200 OK** - **201 Created** - From POSTs that create something new. - **204 No Content** - When there is no answer, ex: \"/:anAction\" or DELETE. - **400 Bad request** -   - **4000** Bad Request Exception   - **11000** Illegal Filter Exception   - **12000** Path Param Exception   - **24000**   Cryptography Exception - **401 Unauthorized** - When authentication is required and has failed or has not yet been provided   -  **3000** Authentication Exception   -  **9000** Security Exception - **403 Forbidden** - When AuthorisationManager says no. - **404 Not Found** - For content/IDs that does not exist.   -  **6000** Not Found Exception - **409 Conflict** - Such as an edit conflict between multiple simultaneous updates   -  **7000** Object Exists Exception   -  **8000** Revision Exception   - **10000** Locked Exception   - **14000** Duplicate entry - **422 Bad Request** - For Required fields or things like malformed payload.   - **15000** Value Validation Exception   - **16000** Mapping Exception   - **17000** Sorting Exception   - **18000** Validation Exception   - **21000** Param Exception   - **22000** Invalid JSON Exception   - **23000**   Result Set Too Large Exception - **429 Too Many Requests** - Request rate limit hit - **500 Internal Error** -  Unexpected condition was encountered and no more specific message is suitable   -  **1000** Exception
  *
- * OpenAPI spec version: 2.33.1
+ * OpenAPI spec version: 2.33.2
  * 
  * Generated by: https://github.com/swagger-api/swagger-codegen.git
- * Swagger Codegen version: 3.0.5
+ * Swagger Codegen version: 2.4.2
  */
+
 /**
  * NOTE: This class is auto generated by the swagger code generator program.
  * https://github.com/swagger-api/swagger-codegen
@@ -57,22 +58,23 @@ class TimesheetEntry implements ModelInterface, ArrayAccess
       */
     protected static $swaggerTypes = [
         'id' => 'int',
-'version' => 'int',
-'changes' => '\Tripletex\Model\Change[]',
-'url' => 'string',
-'project' => '\Tripletex\Model\Project',
-'activity' => '\Tripletex\Model\Activity',
-'date' => 'string',
-'hours' => 'double',
-'chargeable_hours' => 'double',
-'employee' => '\Tripletex\Model\Employee',
-'time_clocks' => '\Tripletex\Model\TimeClock[]',
-'comment' => 'string',
-'locked' => 'bool',
-'chargeable' => 'bool',
-'invoice' => '\Tripletex\Model\Invoice',
-'hourly_rate' => 'BigDecimal',
-'hourly_cost' => 'BigDecimal'    ];
+        'version' => 'int',
+        'changes' => '\Tripletex\Model\Change[]',
+        'url' => 'string',
+        'project' => '\Tripletex\Model\Project',
+        'activity' => '\Tripletex\Model\Activity',
+        'date' => 'string',
+        'hours' => 'double',
+        'chargeable_hours' => 'double',
+        'employee' => '\Tripletex\Model\Employee',
+        'time_clocks' => '\Tripletex\Model\TimeClock[]',
+        'comment' => 'string',
+        'locked' => 'bool',
+        'chargeable' => 'bool',
+        'invoice' => '\Tripletex\Model\Invoice',
+        'hourly_rate' => 'float',
+        'hourly_cost' => 'float'
+    ];
 
     /**
       * Array of property to format mappings. Used for (de)serialization
@@ -81,22 +83,23 @@ class TimesheetEntry implements ModelInterface, ArrayAccess
       */
     protected static $swaggerFormats = [
         'id' => 'int32',
-'version' => 'int32',
-'changes' => null,
-'url' => null,
-'project' => null,
-'activity' => null,
-'date' => null,
-'hours' => 'double',
-'chargeable_hours' => 'double',
-'employee' => null,
-'time_clocks' => null,
-'comment' => null,
-'locked' => null,
-'chargeable' => null,
-'invoice' => null,
-'hourly_rate' => null,
-'hourly_cost' => null    ];
+        'version' => 'int32',
+        'changes' => null,
+        'url' => null,
+        'project' => null,
+        'activity' => null,
+        'date' => null,
+        'hours' => 'double',
+        'chargeable_hours' => 'double',
+        'employee' => null,
+        'time_clocks' => null,
+        'comment' => null,
+        'locked' => null,
+        'chargeable' => null,
+        'invoice' => null,
+        'hourly_rate' => null,
+        'hourly_cost' => null
+    ];
 
     /**
      * Array of property to type mappings. Used for (de)serialization
@@ -126,22 +129,23 @@ class TimesheetEntry implements ModelInterface, ArrayAccess
      */
     protected static $attributeMap = [
         'id' => 'id',
-'version' => 'version',
-'changes' => 'changes',
-'url' => 'url',
-'project' => 'project',
-'activity' => 'activity',
-'date' => 'date',
-'hours' => 'hours',
-'chargeable_hours' => 'chargeableHours',
-'employee' => 'employee',
-'time_clocks' => 'timeClocks',
-'comment' => 'comment',
-'locked' => 'locked',
-'chargeable' => 'chargeable',
-'invoice' => 'invoice',
-'hourly_rate' => 'hourlyRate',
-'hourly_cost' => 'hourlyCost'    ];
+        'version' => 'version',
+        'changes' => 'changes',
+        'url' => 'url',
+        'project' => 'project',
+        'activity' => 'activity',
+        'date' => 'date',
+        'hours' => 'hours',
+        'chargeable_hours' => 'chargeableHours',
+        'employee' => 'employee',
+        'time_clocks' => 'timeClocks',
+        'comment' => 'comment',
+        'locked' => 'locked',
+        'chargeable' => 'chargeable',
+        'invoice' => 'invoice',
+        'hourly_rate' => 'hourlyRate',
+        'hourly_cost' => 'hourlyCost'
+    ];
 
     /**
      * Array of attributes to setter functions (for deserialization of responses)
@@ -150,22 +154,23 @@ class TimesheetEntry implements ModelInterface, ArrayAccess
      */
     protected static $setters = [
         'id' => 'setId',
-'version' => 'setVersion',
-'changes' => 'setChanges',
-'url' => 'setUrl',
-'project' => 'setProject',
-'activity' => 'setActivity',
-'date' => 'setDate',
-'hours' => 'setHours',
-'chargeable_hours' => 'setChargeableHours',
-'employee' => 'setEmployee',
-'time_clocks' => 'setTimeClocks',
-'comment' => 'setComment',
-'locked' => 'setLocked',
-'chargeable' => 'setChargeable',
-'invoice' => 'setInvoice',
-'hourly_rate' => 'setHourlyRate',
-'hourly_cost' => 'setHourlyCost'    ];
+        'version' => 'setVersion',
+        'changes' => 'setChanges',
+        'url' => 'setUrl',
+        'project' => 'setProject',
+        'activity' => 'setActivity',
+        'date' => 'setDate',
+        'hours' => 'setHours',
+        'chargeable_hours' => 'setChargeableHours',
+        'employee' => 'setEmployee',
+        'time_clocks' => 'setTimeClocks',
+        'comment' => 'setComment',
+        'locked' => 'setLocked',
+        'chargeable' => 'setChargeable',
+        'invoice' => 'setInvoice',
+        'hourly_rate' => 'setHourlyRate',
+        'hourly_cost' => 'setHourlyCost'
+    ];
 
     /**
      * Array of attributes to getter functions (for serialization of requests)
@@ -174,22 +179,23 @@ class TimesheetEntry implements ModelInterface, ArrayAccess
      */
     protected static $getters = [
         'id' => 'getId',
-'version' => 'getVersion',
-'changes' => 'getChanges',
-'url' => 'getUrl',
-'project' => 'getProject',
-'activity' => 'getActivity',
-'date' => 'getDate',
-'hours' => 'getHours',
-'chargeable_hours' => 'getChargeableHours',
-'employee' => 'getEmployee',
-'time_clocks' => 'getTimeClocks',
-'comment' => 'getComment',
-'locked' => 'getLocked',
-'chargeable' => 'getChargeable',
-'invoice' => 'getInvoice',
-'hourly_rate' => 'getHourlyRate',
-'hourly_cost' => 'getHourlyCost'    ];
+        'version' => 'getVersion',
+        'changes' => 'getChanges',
+        'url' => 'getUrl',
+        'project' => 'getProject',
+        'activity' => 'getActivity',
+        'date' => 'getDate',
+        'hours' => 'getHours',
+        'chargeable_hours' => 'getChargeableHours',
+        'employee' => 'getEmployee',
+        'time_clocks' => 'getTimeClocks',
+        'comment' => 'getComment',
+        'locked' => 'getLocked',
+        'chargeable' => 'getChargeable',
+        'invoice' => 'getInvoice',
+        'hourly_rate' => 'getHourlyRate',
+        'hourly_cost' => 'getHourlyCost'
+    ];
 
     /**
      * Array of attributes where the key is the local name,
@@ -231,6 +237,8 @@ class TimesheetEntry implements ModelInterface, ArrayAccess
     {
         return self::$swaggerModelName;
     }
+
+    
 
     
 
@@ -670,7 +678,7 @@ class TimesheetEntry implements ModelInterface, ArrayAccess
     /**
      * Gets hourly_rate
      *
-     * @return BigDecimal
+     * @return float
      */
     public function getHourlyRate()
     {
@@ -680,7 +688,7 @@ class TimesheetEntry implements ModelInterface, ArrayAccess
     /**
      * Sets hourly_rate
      *
-     * @param BigDecimal $hourly_rate hourly_rate
+     * @param float $hourly_rate hourly_rate
      *
      * @return $this
      */
@@ -694,7 +702,7 @@ class TimesheetEntry implements ModelInterface, ArrayAccess
     /**
      * Gets hourly_cost
      *
-     * @return BigDecimal
+     * @return float
      */
     public function getHourlyCost()
     {
@@ -704,7 +712,7 @@ class TimesheetEntry implements ModelInterface, ArrayAccess
     /**
      * Sets hourly_cost
      *
-     * @param BigDecimal $hourly_cost hourly_cost
+     * @param float $hourly_cost hourly_cost
      *
      * @return $this
      */
@@ -784,3 +792,5 @@ class TimesheetEntry implements ModelInterface, ArrayAccess
         return json_encode(ObjectSerializer::sanitizeForSerialization($this));
     }
 }
+
+
